+++ src/Serilog.Sinks.AzureBlobStorage/Sinks/AzureBlobProvider/DefaultCloudBlobProvider.cs
@@ -17,8 +17,9 @@ using System;
 using System.IO;
 using System.Net;
 using System.Threading.Tasks;
-using Microsoft.WindowsAzure.Storage;
-using Microsoft.WindowsAzure.Storage.Blob;
+using Microsoft.Azure.Storage;
+using Microsoft.Azure.Storage.Blob;
+
 
 namespace Serilog.Sinks.AzureBlobStorage.AzureBlobProvider
 {