+++ src/Serilog.Sinks.AzureBlobStorage/Sinks/AzureBlobProvider/ICloudBlobProvider.cs
@@ -14,8 +14,8 @@
 // limitations under the License.
 
 using System.Threading.Tasks;
-using Microsoft.WindowsAzure.Storage;
-using Microsoft.WindowsAzure.Storage.Blob;
+using Microsoft.Azure.Storage;
+using Microsoft.Azure.Storage.Blob;
 
 namespace Serilog.Sinks.AzureBlobStorage.AzureBlobProvider
 {