+++ src/Serilog.Sinks.AzureBlobStorage/Sinks/AzureBlobStorage/AzureBlobStorageSink.cs
@@ -14,7 +14,7 @@
 // limitations under the License.
 
 using System.Threading;
-using Microsoft.WindowsAzure.Storage;
+using Microsoft.Azure.Storage;
 using Serilog.Core;
 using Serilog.Events;
 using Serilog.Formatting;