+++ src/Serilog.Sinks.AzureBlobStorage/Sinks/AzureBlobStorage/DefaultAppendBlobBlockWriter.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using System.Threading.Tasks;
-using Microsoft.WindowsAzure.Storage;
-using Microsoft.WindowsAzure.Storage.Blob;
+using Microsoft.Azure.Storage;
+using Microsoft.Azure.Storage.Blob;
 
 namespace Serilog.Sinks.AzureBlobStorage
 {