+++ src/Serilog.Sinks.AzureBlobStorage/Sinks/AzureBlobStorage/IAppendBlobBlockWriter.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using Microsoft.WindowsAzure.Storage.Blob;
+using Microsoft.Azure.Storage.Blob;
 
 namespace Serilog.Sinks.AzureBlobStorage
 {