+++ tests/Serilog.Sinks.AzureBlobStorage.UnitTest/LoggingUT.cs
@@ -1,6 +1,6 @@
 using System;
-using Microsoft.WindowsAzure.Storage;
-using Microsoft.WindowsAzure.Storage.Blob;
+using Microsoft.Azure.Storage;
+using Microsoft.Azure.Storage.Blob;
 using Xunit;
 
 namespace Serilog.Sinks.AzureBlobStorage.UnitTest